// src/types.ts
enum Tone {
    C = 'C',
    Cs = 'C#',
    D = 'D',
    Ds = 'D#',
    E = 'E',
    F = 'F',
    Fs = 'F#',
    G = 'G',
    Gs = 'G#',
    A = 'A',
    As = 'A#',
    B = 'B'
}

const ALL_TONES = [Tone.C, Tone.Cs, Tone.D, Tone.Ds, Tone.E, Tone.F, Tone.Fs, Tone.G, Tone.Gs, Tone.A, Tone.As, Tone.B];

export {
    ALL_TONES,
    Tone
}